parsers: removed callback from findEndOfCueBlock
